Split `compile/axis` into `compile/axis/*`
